requirements.txt
feedparser
numpy
pydub
num2words
google-cloud-texttospeech
google-cloud-translate
google-cloud-aiplatform
google-generativeai
pyloudnorm
rapidfuzz
streamlit
wordcloud
matplotlib
